\name{RFORNAT-package}
\alias{RFORNAT-package}
\alias{RFORNAT}
\docType{package}
\title{
\packageTitle{RFORNAT}
}
\description{
\packageDescription{RFORNAT}
}
\details{

\packageDESCRIPTION{RFORNAT}
\packageIndices{RFORNAT}
}
\author{
\packageAuthor{RFORNAT}
\packageMaintainer{RFORNAT}
}
\references{
}
\keyword{}
\seealso{
\code{}
}
\examples{}
