\name{RFORNAT-package}
\alias{RFORNAT-package}
\alias{RFORNAT}
\docType{package}
\title{
\packageTitle{RFORNAT}
}
\description{
\packageDescription{RFORNAT is an R package to analyze the FORNAT fish database}
}
\details{

\packageDESCRIPTION{RFORNAT is an R package to analyze the FORNAT fish database}
\packageIndices{RFORNAT}
Currently, it contains the function DatenbankAuswertung()
}
\author{
\packageAuthor{DFR}

Maintainer: \packageMaintainer{DFR}
}
\examples{
DatenbankAuswertung(befischung=678,input=database.xlsx)
}
