\name{RFORNAT-package}
\alias{RFORNAT-package}
\alias{RFORNAT}
\docType{package}
\title{
\packageTitle{RFORNAT}
}
\description{
\packageDescription{RFORNAT}
}
\details{

\packageDESCRIPTION{RFORNAT}
\packageIndices{RFORNAT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RFORNAT}

Maintainer: \packageMaintainer{RFORNAT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
